Add: logic and erase modal
