+++ src/components/Skeleton/index.tsx
@@ -1,5 +1,4 @@
-import { ReactNode, useCallback, useEffect, useMemo, useState } from "react";
-import { useCookies } from "react-cookie";
+import { ReactNode } from "react";
 import { useLocation } from "react-router-dom";
 
 import {
@@ -10,16 +9,12 @@ import {
 import HeaderBar from "components/HeaderBar";
 import Loading from "components/Loading";
 import { ModalTerms } from "components/ModalPopup";
-import ModalSuggestApp from "components/ModalPopup/ModalSuggestApp";
 import Error from "pages/Error";
 
 import Navigation from "./Navigation";
 
 import errorAtom from "atoms/error";
-import isAppAtom from "atoms/isApp";
 import { useRecoilValue } from "recoil";
-
-import isMobile from "tools/isMobile";
 
 type ContainerProps = {
   children: ReactNode;
@@ -44,44 +39,13 @@ const Container = ({ children }: ContainerProps) => (
 );
 
 const Skeleton = ({ children }: SkeletonProps) => {
-  const {
-    id: userId,
-    agreeOnTermsOfService: isAgreeOnTermsOfService,
-    deviceType,
-  } = useValueRecoilState("loginInfo") || {};
+  const { id: userId, agreeOnTermsOfService: isAgreeOnTermsOfService } =
+    useValueRecoilState("loginInfo") || {};
   const error = useRecoilValue(errorAtom);
   const isLoading = userId === null;
 
   const location = useLocation();
   const { pathname } = location;
-
-  const [cookies, setCookies] = useCookies(["isOpposeSuggestApp"]);
-  const isOpposeSuggestApp = !!cookies?.isOpposeSuggestApp;
-  const isApp = useRecoilValue(isAppAtom) || deviceType === "app";
-  const [isAndroid, isIOS] = isMobile();
-  const [isTryCloseSuggestApp, setIsTryCloseSuggestApp] = useState(false);
-  const isSuggestApp = useMemo(
-    () =>
-      (isAndroid || isIOS) &&
-      !isApp &&
-      !isOpposeSuggestApp &&
-      !isTryCloseSuggestApp,
-    [isAndroid, isIOS, isApp, isOpposeSuggestApp, isTryCloseSuggestApp]
-  );
-  const setIsOpenSuggestApp = useCallback(
-    () => setIsTryCloseSuggestApp(true),
-    []
-  );
-
-  // 앱 웹뷰일 경우, 앱 설치 유도 팝업 띄우기를 중단합니다 그리고 쿠키를 설정합니다.
-  useEffect(() => {
-    if (isApp && !isOpposeSuggestApp) {
-      const expirationDate = new Date();
-      expirationDate.setFullYear(expirationDate.getFullYear() + 10);
-      setCookies("isOpposeSuggestApp", true, { expires: expirationDate });
-      setIsTryCloseSuggestApp(true);
-    }
-  }, [isApp, isOpposeSuggestApp]);
 
   // loginIngo, taxiLocations, myRooms, notificationOptions 초기화 및 동기화
   useSyncRecoilStateEffect();
@@ -120,10 +84,6 @@ const Skeleton = ({ children }: SkeletonProps) => {
       <Navigation />
       <HeaderBar />
       {children}
-      <ModalSuggestApp
-        isOpen={isSuggestApp}
-        onChangeIsOpen={setIsOpenSuggestApp}
-      />
       <ModalTerms isOpen={!!userId && !isAgreeOnTermsOfService} />
       <div css={{ height: "88px" }} />
     </Container>
