+++ src/pages/Home/SuggestAppTopBar.tsx
@@ -1,8 +1,14 @@
+import { useLocation } from "react-router-dom";
+
 import theme from "tools/theme";
+import { getDynamicLink } from "tools/trans";
 
 import { ReactComponent as TaxiIcon } from "static/assets/TaxiAppIcon.svg";
 
 const SuggestAppTopBar = () => {
+  const { pathname, search } = useLocation();
+  const currentPath = pathname + search;
+
   return (
     <div
       css={{
@@ -10,14 +16,21 @@ const SuggestAppTopBar = () => {
         flexDirection: "row",
         alignItems: "center",
         justifyContent: "space-between",
-        width: "100%",
+        // width: "100%",
         height: "50px",
         backgroundColor: theme.purple,
 
         padding: "0 10px",
       }}
     >
-      <div css={{ display: "flex" }}>
+      <div
+        css={{
+          display: "flex",
+          gap: "10px",
+          alignItems: "center",
+          justifyContent: "center",
+        }}
+      >
         <TaxiIcon css={{ width: "30px", height: "30px" }} />
         <div>
           <div css={{ ...theme.font14_bold, color: theme.white }}>
@@ -28,7 +41,21 @@ const SuggestAppTopBar = () => {
           </div>
         </div>
       </div>
-      <div>열기</div>
+      <a
+        href={getDynamicLink(currentPath, false)}
+        target="_blank"
+        rel="noreferrer"
+        css={{
+          ...theme.font14_bold,
+          borderRadius: "15px",
+          padding: "7px 20px",
+          backgroundColor: theme.white,
+          boxShadow: theme.shadow_3_up,
+          textDecoration: "none",
+        }}
+      >
+        열기
+      </a>
     </div>
   );
 };
